+++ config/bootstrap_plugins.php
@@ -24,6 +24,8 @@ if (file_exists(PLUGINS . DS . 'Passbolt' . DS . 'RememberMe')) {
 }
 
 // Add tags plugin if present.
-if (file_exists(PLUGINS . DS . 'Passbolt' . DS . 'WebInstaller')) {
-    Plugin::load('Passbolt/WebInstaller', ['bootstrap' => true, 'routes' => true]);
+if (defined('PASSBOLT_IS_CONFIGURED') && !PASSBOLT_IS_CONFIGURED) {
+    if (file_exists(PLUGINS . DS . 'Passbolt' . DS . 'WebInstaller')) {
+        Plugin::load('Passbolt/WebInstaller', ['bootstrap' => true, 'routes' => true]);
+    }
 }