+++ plugins/Passbolt/WebInstaller/config/routes.php
@@ -18,6 +18,7 @@ use Cake\Routing\Router;
  * Selenium tests routes
  */
 Router::plugin('Passbolt/WebInstaller', ['path' => '/install'], function ($routes) {
+    $routes->setExtensions(['json']);
 
     $routes->connect('/', ['controller' => 'SystemCheck', 'action' => 'index'])
         ->setMethods(['GET']);
@@ -50,8 +51,5 @@ Router::plugin('Passbolt/WebInstaller', ['path' => '/install'], function ($route
         ->setMethods(['GET']);
 
     $routes->connect('/account_creation', ['controller' => 'AccountCreation', 'action' => 'index'])
-        ->setMethods(['GET', 'POST']);
-
-    $routes->connect('/complete', ['controller' => 'Installation', 'action' => 'complete'])
         ->setMethods(['GET', 'POST']);
 });