+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 ## [Unreleased]
 ### Added
-- PASSBOLT-2694: Implement the webinstaller feature
+- PASSBOLT-2694: Implement the Web Installer feature
+
+### Improved
+- PASSBOLT-3119: The Web Installer should control the route with a middleware
+- PASSBOLT-3153: The Web Installer healtchecks should ensure the config files can be written before continuing
+- PASSBOLT-3120: Improve the Web Installer code coverage
+- PASSBOLT-3127: The Web Installer should change the config folder permissions after the installation is completed
+
+### Fixed
 - PASSBOLT-2963: Persist the license in a file